Expose use error
