+++ src/index.ts
@@ -19,6 +19,7 @@ import {
   useSubscription,
   useResultCache,
   useMeta,
+  useError,
   RestProvider,
   NetworkErrorBoundary,
   NetworkError,
@@ -27,23 +28,23 @@ import { Request as RequestType } from 'superagent';
 import { AbstractInstanceType, RequestOptions, Method } from './types';
 
 export type DeleteShape<
-S extends schemas.Entity,
-Params extends Readonly<object> = Readonly<object>
+  S extends schemas.Entity,
+  Params extends Readonly<object> = Readonly<object>
 > = DeleteShape<S, Params>;
 export type MutateShape<
-S extends Schema,
-Params extends Readonly<object> = Readonly<object>,
-Body extends Readonly<object> | void = Readonly<object> | undefined
+  S extends Schema,
+  Params extends Readonly<object> = Readonly<object>,
+  Body extends Readonly<object> | void = Readonly<object> | undefined
 > = MutateShape<S, Params, Body>;
 export type ReadShape<
-S extends Schema,
-Params extends Readonly<object> = Readonly<object>,
-Body extends Readonly<object> | void = Readonly<object> | undefined
+  S extends Schema,
+  Params extends Readonly<object> = Readonly<object>,
+  Body extends Readonly<object> | void = Readonly<object> | undefined
 > = ReadShape<S, Params, Body>;
 export type RequestShape<
-S extends Schema,
-Params extends Readonly<object> = Readonly<object>,
-Body extends Readonly<object> | void = Readonly<object> | undefined
+  S extends Schema,
+  Params extends Readonly<object> = Readonly<object>,
+  Body extends Readonly<object> | void = Readonly<object> | undefined
 > = RequestShape<S, Params, Body>;
 
 export type Schema<T = any> = Schema<T>;
@@ -67,6 +68,7 @@ export {
   useSubscription,
   useResultCache,
   useMeta,
+  useError,
   NetworkManager,
   NetworkErrorBoundary,
   schemas,
