+++ src/react-integration/hooks/index.ts
@@ -5,10 +5,12 @@ import useRetrieve from './useRetrieve';
 import useResource from './useResource';
 import useSubscription from './useSubscription';
 import useMeta from './useMeta';
+import useError from './useError';
 
 export {
   useFetcher,
   useCache,
+  useError,
   useResultCache,
   useRetrieve,
   useResource,
